refactor(magic): create singlestoredb.magics module
