--- run_personal/__init__.py
@@ -1,4 +1,0 @@
-from run_personal.magic import load_ipython_extension
-
-
-__all__ = ['load_ipython_extension']
