--- run_shared/__init__.py
@@ -1,4 +1,0 @@
-from run_shared.magic import load_ipython_extension
-
-
-__all__ = ['load_ipython_extension']
